+++ tests/test_hw13_ml_computer_vision.cpp
@@ -36,7 +36,7 @@ BOOST_AUTO_TEST_CASE(MlpClassifier_predict) {
 	int label_index = 0;
 	for (const auto &x : x_test) {
 		auto res = ml.predict(x);
-		if (res == y_test[label_index++]) {
+		if (res == static_cast<size_t>(y_test[label_index++])) {
 			correct_predictions++;
 		}
 	}
